Cleanup birb_db.cpp
