+++ include/Database.hpp
@@ -3,7 +3,11 @@
 #include <unordered_map>
 #include <vector>
 
+constexpr char BIRB_DB_PATH[]       	= "/var/lib/birb/birb_db";
+constexpr char BIRB_FAKEROOT_PATH[] 	= "/var/db/fakeroot";
+constexpr char BIRB_PKG_PATH[]      	= "/var/db/pkg";
 constexpr char PKG_SOURCE_CONFIG_PATH[] = "/etc/birb-sources.conf";
+constexpr int DB_LINE_COLUMN_COUNT = 2;
 
 struct pkg_source
 {
@@ -33,8 +37,13 @@ namespace birb
 	/* Returns the raw birb_db file if it exists */
 	std::vector<std::string> read_birb_db();
 
+	std::vector<std::string> find_db_entry(const std::vector<std::string>& db_file, const std::string& pkg_name);
+
 	/* Get the list of installed package names with the birb_db file */
 	std::vector<std::string> get_installed_packages();
+
+	/* Get versions for all packages taking the repository order into account */
+	std::unordered_map<std::string, std::string> get_repo_versions();
 
 	/* Caching */
 	inline std::unordered_map<std::string, std::string> var_cache;
